Add #run_ruby and #run_script
The method #ruby_run executes external Ruby scripts with a separate process. This method allows to specify an input as well as to return the status and changing the directory.

The method #run_script accepts a String or IO with a Ruby script to be executed in a separate process. It allows to return the status and to change the directory.